NetworkAPI commit (#10)
* NetworkAPI commit

This commit holds my NetworkAPI prototype, consisting of various request/response classes, helper enums, and the like.

* Move the status checks from last checkpoint

This was throwing an error for me so I could not merge with main. I passed that test, of course, before submitting checkpoint 1 so I dont think its necessary to run as a test anymore.