Checkpoint5.1 (#30)
* Added basic validation of params, and began exception handling

Did not yet add validation for the file IO helper methods in the process api. I felt this would be very closely tied to handling exceptions and would require the most in-depth validation, so I just wanted to wait for a bit. I also added validation within the custom request classes that get passed as params to public api methods.

* ensured all exceptions are handled gracefully

Added really nice error handling to ensure all exceptions that could happen, both expected and unexpected, are handled properly and described to the user in some cases. Also added a bit of validation to the private file helper methods in the process api. Specifically, just checking if the files exist before trying to read or write. Also added the checkpoint5testsuite to the project.

* Added tests

Created tests to check the validation of parameters and also to check that exceptions are handled gracefully.

* fixed checkstyle errors

fixed checkstyle issues

* more small checkstyle stuff

checkstyle issues fixed

* Addresses comments

Addressed reviewer feedback by adding more comments around the catch blocks