Checkpoint6 (#31)
* Added the necessary files

Did the todos in testmultiuser and added the files i will need to proceed

* did testuser

added basic code just to build a request using the input and output files and running the .compute method with that

* Finished checkpoint 6

made soem small changes to process api and network api. also made the multithreaded implementation and changed a couple things to ensure it was working correct. Updated readme with number of cores.

* super weird error

for some reason checkpoint3testsuite is throwing a super strange error, im thinking it may be from adding the new multithreaded implementation in the same package as the other impelementations

* fix checkstyle

* checkstyle