Checkpoint7 (#32)
* Added the necessary files

Did the todos in testmultiuser and added the files i will need to proceed

* did testuser

added basic code just to build a request using the input and output files and running the .compute method with that

* Finished checkpoint 6

made soem small changes to process api and network api. also made the multithreaded implementation and changed a couple things to ensure it was working correct. Updated readme with number of cores.

* super weird error

for some reason checkpoint3testsuite is throwing a super strange error, im thinking it may be from adding the new multithreaded implementation in the same package as the other impelementations

* fix checkstyle

* checkstyle

* Added NetworkProto

Got proto2 all set up, changed build, gitignore, gradle.yml. Added a .proto file for the network api, specifically the .compute method.

* Added Server and Client

Created a grpc server for the netwokr api. created a quick client code to perform computaiton and prompt for inputs. tested to ensure everything is working. Also created teh NetworkService IMplementation

* Final commit, implemented process API grpc changed networkAPI to use the new grpc

Ran into a super weird error in using | as a delimiter, forgot it was a special character so ended up having to debug that. Did testing, though and everything seems to be working great.

* checkstyle fixed

* compile error

so weird, a test a fialing on github but passes locally, something weird with file io.

* trying to fix weird error

* trying to fix compile error

* same as before

* one last try

if this doesnt work im moving these into a non src folder and moving on, they all pass locally

* Had to move out of src folder

no idea whats causing the issue and no idea how to fix it, tried everything, tests pass locally