Checkpoint8.1 (#33)
* Added performance testing infra

Created a timing logging util and wrapped varrious sections of the multithreaded compute method in it. Will use it to help find perfoamcne issues and for testing later on.

* Fixed perfomancec test and multithreaded network api

had a compilation issue in these two files, just fixed them real fast

* totalyl messed up hte perofmance capture

totally forgot to wrap the actual computation call in its own section for perfomance testing. this is obviously a vital piece in terms of performance. just updated the compute() method so we can measure the performComputation() calls

* Added reset method to logger

since this is a fully static class we will need to be able to reset the TimingLogger once we want to compared speeds of the old and improved code.