Improved computation performance and added benchmark test
made the computation way faster, added a new method both to multithreaded network api and conceptualAPI that uses the new fast computatiion. Created a test the tracks the average time the computation takes for the new and old versions, then computes the % difference and passes if that % difference is larger than 10.