Checkpoint8.2 (#34)
* Improved computation performance and added benchmark test

made the computation way faster, added a new method both to multithreaded network api and conceptualAPI that uses the new fast computatiion. Created a test the tracks the average time the computation takes for the new and old versions, then computes the % difference and passes if that % difference is larger than 10.

* Added separate performance comparison and updated readme with performance improvment

Added performance comparison to get the exact % improvements between the original and new computation in addition to the test to check if performance improvement is 10% or larger. Also updated readme to now detail the improvements that were made and the process I used to find them.

* Addressed checkpoint 7 feedback

Fixed an issue with my ProcessAPI that led to errors when tryign to run the NetworkClient using in memory inputs. (Not sure how i missed this issue initally, very poor testing on my part). But it turned out I was returning a null message in hte LoadResponse for CUSTOM resources in my ProcessAPI which was cuasing a null pointer when we tried to access that message field. I was returning a non-null response for loading File inputs so thats why it was only happening for in memory inputs. I just changed the regular process api to return an actual message for sucessfully loading custom resources and that fixed the isssue.